[Issue-2617] Add chain info test script
